Killcode Processor
This processes any killcodes entered into it to shut down the malware. Theres a few hardcoded ones, and it also attempts to read the killcode generated by the main process.

it is always at the front with no controls to delete it.

This contains the already built binary as it is not, in its self. malicious.

It could be edited to process bitcoin payment.